Moved Img upload to Sidenav
