--- src/app/admin/admin.component.html
@@ -1,30 +1,0 @@
-<mat-grid-list cols="3" rowHeight="100px">
-  <mat-grid-tile></mat-grid-tile>
-  <mat-grid-tile>
-    <h1>Admin Panel</h1>
-  </mat-grid-tile>
-</mat-grid-list>
-
-<mat-grid-list cols="3" rowHeight="200px">
-  <mat-grid-tile></mat-grid-tile>
-  <mat-grid-tile>
-    <mat-card>
-      <div style="text-align: center;">
-        <h2>Options</h2>
-        <button mat-raised-button color="primary" (click)="openUpload(uploadfp)" class="mod-button">
-          Floorplans
-        </button>
-      </div>
-    </mat-card>
-  </mat-grid-tile>  
-</mat-grid-list>
-
-<ng-template #uploadfp>
-    <h1 mat-dialog-title>Upload Preview</h1>
-  <form class="uploadFP">
-    <button mat-raised-button color="primary" (click)="fileInput.click()">Upload Floorplan</button>
-    <input #fileInput type="file" (change)="chooseFiles($event)" style="display:none;" />
-    <br>
-    <button mat-raised-button color="primary" style="margin-top: 10%; float: right;">Submit</button>
-  </form>
-</ng-template>
