--- src/app/admin/admin.component.scss
@@ -1,22 +1,0 @@
-$white: white;
-$black: black;
-
-
-mat-card {
-    margin-top: 5%;
-    width: 90%;
-    background: $white;
-    overflow-x: auto;
-}
-
-h2 {
-    color: $black;
-}
-
-button {
-    position: center;
-}
-
-.mod-button {
-    width: 85%;
-}
