+++ src/app/plotting/sidenavmenu/sidenavmenu.component.scss
@@ -23,6 +23,10 @@ mat-sidenav-container {
   width: 100vw;
 }
 
+.mod-button {
+  width: 100%;
+}
+
 .mat-button {
   min-width: 1%;
 }
